+++ renderer/native/android/src/main/java/com/tencent/renderer/FrameworkProxy.java
@@ -18,7 +18,6 @@ package com.tencent.renderer;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
-
 import com.tencent.renderer.component.image.ImageDecoderAdapter;
 import com.tencent.renderer.component.text.FontAdapter;
 import com.tencent.vfs.VfsManager;