+++ renderer/tdf/common/include/renderer/tdf/tdf_render_manager.h
@@ -25,6 +25,8 @@
 #include "dom/dom_node.h"
 #include "dom/render_manager.h"
 #include "footstone/serializer.h"
+#include "vfs/uri_loader.h"
+
 #pragma clang diagnostic push
 #pragma clang diagnostic ignored "-Wdeprecated-copy"
 #pragma clang diagnostic ignored "-Wextra-semi"
@@ -33,60 +35,83 @@
 #include "renderer/tdf/viewnode/root_view_node.h"
 
 namespace hippy {
-inline namespace dom {
+inline namespace render {
+inline namespace tdf {
 
-static std::map<std::string, hippy::render::tdf::ViewNode::node_creator> node_creator_tables_;
-static std::unordered_map<uint32_t, hippy::render::tdf::RootViewNode::UriDataGetter> uri_data_getter_map_;
+static std::map<std::string, ViewNode::node_creator> node_creator_tables_;
 
 void InitNodeCreator();
-void RegisterNodeCreator(const std::string&, const hippy::render::tdf::ViewNode::node_creator&);
-hippy::render::tdf::ViewNode::node_creator GetNodeCreator(const std::string&);
+void RegisterNodeCreator(const std::string &, const ViewNode::node_creator &);
+ViewNode::node_creator GetNodeCreator(const std::string &);
 
-class TDFRenderManager : public RenderManager, public std::enable_shared_from_this<TDFRenderManager> {
-  using RootViewNode = hippy::render::tdf::RootViewNode;
-  using ViewNode = hippy::render::tdf::ViewNode;
+class TDFRenderManager
+    : public RenderManager, public std::enable_shared_from_this<TDFRenderManager> {
 
  public:
-  static footstone::utils::PersistentObjectMap<uint32_t, std::shared_ptr<TDFRenderManager>>& PersistentMap() {
+  static footstone::utils::PersistentObjectMap<uint32_t, std::shared_ptr<TDFRenderManager>> &
+  PersistentMap() {
     return persistent_map_;
   }
 
   TDFRenderManager();
 
-  static void SetUriDataGetter(uint32_t render_id, RootViewNode::UriDataGetter uriDataGetter);
-  RootViewNode::UriDataGetter GetUriDataGetter();
+  uint32_t GetId() { return id_; }
 
-  int32_t GetId() { return id_; }
-
-  void CreateRenderNode(std::weak_ptr<RootNode> root_node, std::vector<std::shared_ptr<DomNode>>&& nodes) override;
-  void UpdateRenderNode(std::weak_ptr<RootNode> root_node, std::vector<std::shared_ptr<DomNode>>&& nodes) override;
-  void MoveRenderNode(std::weak_ptr<RootNode> root_node, std::vector<std::shared_ptr<DomNode>>&& nodes) override;
-  void DeleteRenderNode(std::weak_ptr<RootNode> root_node, std::vector<std::shared_ptr<DomNode>>&& nodes) override;
-  void UpdateLayout(std::weak_ptr<RootNode> root_node, const std::vector<std::shared_ptr<DomNode>>& nodes) override;
-  void MoveRenderNode(std::weak_ptr<RootNode> root_node, std::vector<int32_t>&& moved_ids, int32_t from_pid,
+  void CreateRenderNode(std::weak_ptr<RootNode> root_node,
+                        std::vector<std::shared_ptr<DomNode>> &&nodes) override;
+  void UpdateRenderNode(std::weak_ptr<RootNode> root_node,
+                        std::vector<std::shared_ptr<DomNode>> &&nodes) override;
+  void MoveRenderNode(std::weak_ptr<RootNode> root_node,
+                      std::vector<std::shared_ptr<DomNode>> &&nodes) override;
+  void DeleteRenderNode(std::weak_ptr<RootNode> root_node,
+                        std::vector<std::shared_ptr<DomNode>> &&nodes) override;
+  void UpdateLayout(std::weak_ptr<RootNode> root_node,
+                    const std::vector<std::shared_ptr<DomNode>> &nodes) override;
+  void MoveRenderNode(std::weak_ptr<RootNode> root_node,
+                      std::vector<int32_t> &&moved_ids,
+                      int32_t from_pid,
                       int32_t to_pid) override;
   void EndBatch(std::weak_ptr<RootNode> root_node) override;
   void BeforeLayout(std::weak_ptr<RootNode> root_node) override;
   void AfterLayout(std::weak_ptr<RootNode> root_node) override;
   void AddEventListener(std::weak_ptr<RootNode> root_node, std::weak_ptr<DomNode> dom_node,
-                        const std::string& name) override;
+                        const std::string &name) override;
 
   void RemoveEventListener(std::weak_ptr<RootNode> root_node, std::weak_ptr<DomNode> dom_node,
-                           const std::string& name) override;
-  void CallFunction(std::weak_ptr<RootNode> root_node, std::weak_ptr<DomNode> dom_node, const std::string& name,
-                    const DomArgument& param, uint32_t cb_id) override;
+                           const std::string &name) override;
+  void CallFunction(std::weak_ptr<RootNode> root_node,
+                    std::weak_ptr<DomNode> dom_node,
+                    const std::string &name,
+                    const DomArgument &param,
+                    uint32_t cb_id) override;
 
-  void RegisterShell(uint32_t root_id, const std::shared_ptr<tdfcore::Shell>& shell);
+  void RegisterShell(uint32_t root_id, const std::shared_ptr<tdfcore::Shell> &shell);
 
   void SetDomManager(std::weak_ptr<DomManager> dom_manager) { dom_manager_ = dom_manager; }
-  std::shared_ptr<DomManager> GetDomManager() const { return dom_manager_.lock(); }
+
+  std::shared_ptr<DomManager> GetDomManager() const {
+    return dom_manager_.lock();
+  }
+
+  void SetUriLoader(std::weak_ptr<UriLoader> uri_loader) {
+    FOOTSTONE_CHECK(uri_loader.lock());
+    uri_loader_ = uri_loader;
+  }
+  std::shared_ptr<UriLoader> GetUriLoader() const {
+    FOOTSTONE_CHECK(uri_loader_.lock());
+    return uri_loader_.lock();
+  }
 
  private:
-  footstone::utils::PersistentObjectMap<uint32_t, std::shared_ptr<RootViewNode>> root_view_nodes_map_;
-  int32_t id_;
+  footstone::utils::PersistentObjectMap<uint32_t, std::shared_ptr<RootViewNode>>
+      root_view_nodes_map_;
+  uint32_t id_;
   std::weak_ptr<DomManager> dom_manager_;
-  static inline footstone::utils::PersistentObjectMap<uint32_t, std::shared_ptr<TDFRenderManager>> persistent_map_;
+  std::weak_ptr<UriLoader> uri_loader_;
+  static inline footstone::utils::PersistentObjectMap<uint32_t, std::shared_ptr<TDFRenderManager>>
+      persistent_map_;
 };
 
-}  // namespace dom
+}  // namespace tdf
+}  // namespace render
 }  // namespace hippy